[Fix] Separate Data Model to DiaryViewModel + Fix DiaryDetailView Navigation
