+++ CBLTesting.swiftpm/Assets.xcassets/cc.imageset/Contents.json
@@ -1,15 +1,17 @@
 {
   "images" : [
     {
+      "filename" : "스크린샷 2023-03-23 오후 2.39.22 1.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "스크린샷 2023-03-23 오후 2.39.22.png",
+      "filename" : "스크린샷 2023-03-23 오후 2.39.22 2.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "스크린샷 2023-03-23 오후 2.39.22 3.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
