--- CBLTesting.swiftpm/DataBox.swift
@@ -1,8 +1,0 @@
-//
-//  DataBox.swift
-//  CBLTesting
-//
-//  Created by KimChoonSik on 2023/03/30.
-//
-
-import Foundation
