+++ CBLTesting.swiftpm/EndingView.swift
@@ -1,0 +1,20 @@
+//
+//  EndingView.swift
+//  CBLTesting
+//
+//  Created by Ye Eun Choi on 2023/03/30.
+//
+
+import SwiftUI
+
+struct EndingView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct EndingView_Previews: PreviewProvider {
+    static var previews: some View {
+        EndingView()
+    }
+}
